RAG/requirements.txt
weaviate-client==4.9.3
openai==1.54.3
python-dotenv==1.0.0
click==8.1.7
rich==13.9.4
sentence-transformers==3.3.1
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.10.2
pydantic-settings==2.6.1
httpx
marker-pdf